-add readlink function to the catalog
git-svn-id: https://svn.cern.ch/reps/lcgutil/gfal/branches/gfal_2_0_dev@3792 4525493e-7705-40b1-a816-d608a930855b

+++ src/posix/gfal_posix_local_file.c
@@ -221,7 +221,7 @@ int gfal_local_close(gfal_file_handle fh, GError** err){
 }
 
 ssize_t gfal_local_readlink(const char* path, char* buff, size_t buffsiz, GError** err){
-	const int res = readlink(path + strlen(GFAL_LOCAL_PREFIX), buff, buffsiz);
+	const ssize_t res = readlink(path + strlen(GFAL_LOCAL_PREFIX), buff, buffsiz);
 	if(res <0){
 		gfal_local_report_error(__func__, err);
 	}else
